pyproject.toml
[project]
name = "beshence-vault"
version = "0.0.1"
description = "Beshence Vault"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi-versionizer>=4.0.2",
    "fastapi[standard]>=0.116.1",
    "opaque>=1.0.0",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
    "testcontainers[postgres]>=4.12.0",
]
